refactor: move event types to event module
