+++ crates/ari/src/ari_client/channel/responses.rs
@@ -1,4 +1,3 @@
-use chrono::{DateTime, Utc};
 use derive_getters::Getters;
 use serde::Deserialize;
 
@@ -27,97 +26,6 @@ pub struct RtpStatistics {
 #[serde(rename_all = "snake_case")]
 pub struct ChannelVariable {
     id: String,
-}
-#[derive(Debug, Deserialize, Getters)]
-#[serde(rename_all = "snake_case")]
-pub struct StasisStart {
-    timestamp: DateTime<Utc>,
-    args: Vec<String>,
-    channel: Channel,
-    asterisk_id: String,
-    application: String,
-}
-
-#[derive(Debug, Deserialize, Getters)]
-#[serde(rename_all = "snake_case")]
-pub struct StasisEnd {
-    timestamp: DateTime<Utc>,
-    channel: Channel,
-    asterisk_id: String,
-    application: String,
-}
-
-#[derive(Debug, Deserialize, Getters)]
-#[serde(rename_all = "snake_case")]
-pub struct ChannelCreated {
-    timestamp: DateTime<Utc>,
-    channel: Option<Channel>,
-    asterisk_id: String,
-    application: String,
-}
-
-#[derive(Debug, Deserialize, Getters)]
-#[serde(rename_all = "snake_case")]
-pub struct ChannelDestroyed {
-    timestamp: DateTime<Utc>,
-    cause: i32,
-    cause_txt: String,
-    channel: Channel,
-    asterisk_id: String,
-    application: String,
-}
-
-#[derive(Debug, Deserialize, Getters)]
-#[serde(rename_all = "snake_case")]
-pub struct ChannelVarset {
-    timestamp: DateTime<Utc>,
-    variable: String,
-    value: String,
-    channel: Option<Channel>,
-    asterisk_id: String,
-    application: String,
-}
-
-#[derive(Debug, Deserialize, Getters)]
-#[serde(rename_all = "snake_case")]
-pub struct ChannelHangupRequest {
-    timestamp: DateTime<Utc>,
-    soft: Option<bool>,
-    cause: i32,
-    channel: Channel,
-    asterisk_id: String,
-    application: String,
-}
-
-#[derive(Debug, Deserialize, Getters)]
-#[serde(rename_all = "snake_case")]
-pub struct ChannelDialplan {
-    timestamp: DateTime<Utc>,
-    dialplan_app: String,
-    dialplan_app_data: String,
-    channel: Channel,
-    asterisk_id: String,
-    application: String,
-}
-
-#[derive(Debug, Deserialize, Getters)]
-#[serde(rename_all = "snake_case")]
-pub struct ChannelStateChange {
-    timestamp: DateTime<Utc>,
-    channel: Channel,
-    asterisk_id: String,
-    application: String,
-}
-
-#[derive(Debug, Deserialize, Getters)]
-#[serde(rename_all = "snake_case")]
-pub struct ChannelDtmfReceived {
-    timestamp: DateTime<Utc>,
-    digit: String,
-    duration_ms: i32,
-    channel: Channel,
-    asterisk_id: String,
-    application: String,
 }
 
 #[derive(Debug, Deserialize, Getters)]
